backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "pandas>=2.3.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.24",
    "supabase>=2.17.0",
    "uvicorn[standard]>=0.35.0",
    "openpyxl>=3.1.2",
    "python-multipart>=0.0.6",
]

